feat: home refactor
