+++ client/package.json
@@ -14,8 +14,11 @@
     "@types/node": "^18.11.9",
     "@types/react": "^18.0.25",
     "@types/react-dom": "^18.0.9",
+    "@types/react-router-dom": "^5.3.3",
     "cra-append-sw": "^2.7.0",
     "dotenv": "^16.0.3",
+    "i": "^0.3.7",
+    "npm": "^9.1.2",
     "react": "^18.2.0",
     "react-bootstrap": "^2.6.0",
     "react-datalist-input": "^3.1.0",
